requirements.txt
streamlit
pandas
pathlib
langchain
langchain-community
langchainhub
langgraph
langsmith
sqlalchemy
mysql-connector-python
langchain-experimental
langchain-groq
langchain-openai
plotly
datetime
numpy
tabulate
langchain-anthropic
python-dotenv
ipykernel
anthropic
plotly>=5.0.0
nbformat>=4.2.0
matplotlib